Implement the `pci` command print the C byte array + instruction comments ##disasm (#16792)
