+++ libr/include/r_util/r_print.h
@@ -65,7 +65,6 @@ typedef struct r_print_t {
 	bool scr_prompt;
 	int (*disasm)(void *p, ut64 addr);
 	PrintfCallback oprintf;
-	int interrupt;
 	int big_endian;
 	int width;
 	int limit;
